requirements.txt
# MCP Server
fastmcp
requests
tavily-python

# Environment configuration
python-dotenv

# LangChain and LangGraph
langchain
langchain-openai
langchain-core
langgraph

# MCP Client Adapter
langchain-mcp-adapters

# FastAPI Web App
fastapi
uvicorn[standard]
httpx
jinja2
python-multipart
